fix: keep homepage hotfix within remediation scope
